requirements.txt
# Зависимости для Telegram-бота командной работы

# Основная библиотека для создания Telegram ботов
aiogram==3.13.1

# Библиотека для работы с переменными окружения
python-dotenv==1.0.0

